requirements.txt
Jinja2==3.1.2
lark==1.1.9
matplotlib==3.7.2
numpy==1.25.1
openai==1.30.1
pandas==2.0.3
plotly==5.15.0
python-dotenv==1.0.1
ruamel.base==1.0.0
scikit_learn==1.3.0
scipy==1.13.0
tenacity==8.2.2
torch==2.1.0
torch_xla==2.3.0
tqdm==4.65.0
transformers==4.40.2
